+++ CHANGELOG.md
@@ -109,17 +109,27 @@
 
 - Add class `MLL\Utils\Specification` for logical combinations of predicates
 
+## v6.0.0
+
+### Changed
+
+- Breaking Change: Rename class `MLL\Utils\PHPStan\Rules\VariableNameIdToIDRule` to `MLL\Utils\PHPStan\Rules\NameIdToIDRule`
+- Make `MLL\Utils\PHPStan\Rules\NameIdToIDRule` handle variable names, parameter names, method names, and class names for incorrect capitalization of `Id`
+- Add `Idt` and `Identical` to the list of incorrect capitalizations for `Id` of `MLL\Utils\PHPStan\Rules\NameIdToIDRule`
+- Add RuleIdentifier `mll.capitalizationOfID` to `MLL\Utils\PHPStan\Rules\NameIdToIDRule`
+- Add PHPStan-rule `MLL\Utils\PHPStan\Rules\CanonicalCapitalizations` to check spelling of LabID
+
 ## v5.8.0
 
 ### Added
 
-- Add PHPStan-Rule `MLL\Utils\PHPStan\Rules\ThrowableClassNameRule`
+- Add PHPStan-rule `MLL\Utils\PHPStan\Rules\ThrowableClassNameRule`
 
 ## v5.7.0
 
 ### Added
 
-- Add PHPStan-Rule `MLL\Utils\PHPStan\Rules\VariableNameIdToIDRule`
+- Add PHPStan-rule `MLL\Utils\PHPStan\Rules\VariableNameIdToIDRule`
 
 ## v5.6.0